Test Rails.app.creds
